feat: styles.css, main.js: toggle nav on and off For styles.css the bottom is set to -100%, that the nav isn't showing.
FOR JS File:
We get the getElementById with the id from index.html where there is id="".

For MENU SHOW:
It is still watching if navToggle has an Event like clikc on the button or icon and if that is true, it will add the show-menu tag.

For MENU HIDE:
It is still watching in the background if there is an Event like clicking on the navClose button and if that happens, it will remove the nav.