feat: main.js: toggle Skills open and close This is the full code for toggling the Skills open and close.
First, we get the Elements by getElementsByClassName, that was wrong , for the last time. The last commit I said getElementByClassName, but thats weong, right would be getLementsByClassName.

Then we have a function toggleSkills. That function works with 0 and 1.

The last thing has an EventListener, if we click on a Skill, it toggles 'toggleSkills' and it opens. All the others will close.