feat: index.html, styles.css: added line to connect the dots For CSS: Created a line that begins at the fist dot and stops for the last dot.
For HTML: I command out a class, called: "qualification_line" for the last QUALIFICATION 3, that there is no line behind the last dot. If you want to add more dots, so more QUALIFICATIONS under the Education qualification, or other qualifications, than you have to un-command it and command it for the last one.