feat: index.html: create a service in service section Created the first service in the service section.
On the very top is the icon, then the title Ui/Ux Designer, which can be modified later on, after that comes the button and then the h4 heading. After that the secription of the Service. The services can be added more, there could exist hundrets of them, but that would be very unconvinient