changed max-width fo container and nav Big step: Managed to edit the max-width so there is not so much free space on the right and left side of the screen. So the website was before of the changes much more centered and now the feeling on wide view is much better.
I also added a new screen size of 1280px for even larger devices so there is no margin and the home_blob is also bigger.
For 1024 px: I added the .container for the margin-left auto and modified the tag main (.main) to add a margin-left of 5rem that the text is more in the middle than sticks on the very left side.